Use Jekyller to auto-generate Facets and Views
We're using Thor to create a CLI that runs
[Jekyller](https://github.umn.edu/sdp/jekyller). This generates all of
the markdown files that document Views and Facets.

Because all of these files are put into the `site.pages` collection,
we've replaced the dynamic navigation with a static one. The dynamic
navigation would show all pages, and there were a lot of them.

This is not the final product. We still need to update the readme and
changelog when publishing. And the CLI needs better documentation. But
we can deploy this and show the new pages to the SDP team.