Fix (?) subtree weirdness on publish
Github pages is not ublishing with this error:

 Your site is having problems building: The submodule lib/jekyller/tmp
 was not properly initialized with a .gitmodules file. For more
 information, see
 https://help.github.com/articles/page-build-failed-missing-submodule.

I had accidentally committed a version of jekyller's tmp folder
containing the sdp_reporting repo. That seems to mess with github pages.

So, going back to ignoring jekyller's tmp folder (we do not need it,
when we clone sdp and sdp_reporting, the tmp folder will get built) and
also cleaning up that directory.