Remove jekyller and thor
Moving the thor cli into jekyller, so that the projects that will update
this site can use jekyller to make changes to this project directly.

(For example sdp can subtree this site and jekyller. Then upon deploy,
it can use jekyller to add/update pages to the site and then push.)