Build facets and views index with jekyll template
Switch from building the facets and views index pages as part of
the `Jekyller::Library#document` method to building them with a Liquid
template.

The main reason to do this is to allow different projects to update the
site separately. Both sdp and sdp_reporting have views that get
documented in the views folder. When Jekyller::Library created the index
page, it meant that publishing changes from one project required
checking out the other so that a combined index could be generated.
This made the publishing process and the dependencies harder to manage.

This change means that these projects only need to have an up to date
copy of this jekyll site and they do not have to know who else might be
updating it (or clone them as part of the publishing process).

A side benefit is that it also makes it easy to have the index page in
alphabetical order, since it gets built after all the other pages get
built.